Attachment support and HTML parsing improvements
spyhunter99 reported several problems:
- issues on table conversion with whitespaces
  - Solved by transitioning away from html2text [1]
- lack of attachment support
  - Quickly coded this up as it seems like a very nice to have feature
- inability to export child pages
  - Solved by updating the atlassian package

1) Transitioning from html2text

This was in the pipeline, as Atlassian likes to have its' own format
for HTML I migrated to `markdownify` which seems to be well-maintained,
and is based on BeautifulSoup. Because of this I also wrote a
preprocessor that - in an extremely crude way - replaces the `ac:images`
tags with proper `img` that later can be parsed by markdownify.

This enables attachment embedding / linking as well as future support
for Atlassians myriad of nonstandard HTML tags.